dashboard/requirements.txt
Django==3.0.2
django-webpack-loader==0.6.0
python-decouple==3.1
djangorestframework==3.10.3
xlrd==1.2.0
boto3==1.11.0
django-storages==1.8
Pillow==7.0.0
